optimized facerec

+++ py/apps/videofacerec/main.py
@@ -25,27 +25,37 @@ def vid_parse(vid_dir_path):
 			nametags = initiate_vid(
 				model_param = 'model.pkl',
 				vidpath_param = vid_dir_path+'/'+vid,
-				dataset_param = '/Users/kage/faceimagesDB/facedb')
+				dataset_param = '/Users/kage/faceimagesDB/facedb'
+				)
 			yield [vid,nametags]
 
 
 
 def gener_db(vid_dir_path):
 	VIDDB = {}
+	db ={}
 	nametags = vid_parse(vid_dir_path)
 	for vid,tags in nametags:
 		VIDDB[vid] = tags
-	with open('vidtags.pickle', 'rb') as f:
-	  db = pickle.load(f)
-	  for datum in VIDDB:
-		  db[datum] = VIDDB[vid]
 
-	with open('vidtags.pickle', 'wb') as f:
-		pickle.dump(db, f)
+	if not os.path.exists('vidtags.pickle.db'):
+		for datum in VIDDB:
+			  db[datum] = VIDDB[vid]
+		with open('vidtags.pickle', 'wb') as f:
+			pickle.dump(db, f)
 
+	else:
+		with open('vidtags.pickle', 'rb') as f:
+		  db = pickle.load(f)
+		  for datum in VIDDB:
+			  db[datum] = VIDDB[vid]
 
+		with open('vidtags.pickle', 'wb') as f:
+			pickle.dump(db, f)
 
-gener_db('/Users/kage/xvids/test/')
+	print db
+
+gener_db('/Users/kage/xvids/test')
 
 testdata = {'annie.mp4': {'sara_sigmundsdottir': {'hits': 14, 'time': 47}, 'annie_thorisdottir': {'hits': 49, 'time': 25}, 'mat_fraser': {'hits': 10, 'time': 21}, 'rich_froning': {'hits': 5, 'time': 21}, 'samantha_briggs': {'hits': 0, 'time': 22}}}
